Move endpoint property setup to main setup class
Move the endpoint property definitions from the plugin load
function into the plugin setup class proper. This will allow for
better separation of concerns as well as a more complex users
property definition. We'll need to loop through all of the site
registered users to build the endpoint users array.