Add dedicated method to set the plugin properties
Using a dedicated method we can build an array of users to feed to
the Accessibility API endpoint and in the future can use it to
define any other plugin properties we might need.

This commit also includes some tidying of @todo comments;
redefines the `set_endpoint_props` method as private; and
sanitizes both endpoint properties before saving them.