:fire: Remove WP cron scheduler and replace with `admin_init` hook
No longer using the WP cron scheduler to call the method that
retrieves data from the API. It didn't work well with the method
of storing the API results in a transient, leaving periods when
the transient had expired but the cron job had yet to run.

Instead, Use an `admin_init` action hook to run the method that
retrieves data from the API on every admin page load instead of
once per hour. The request method still only pings the API if the
transient has expired, so we're not increasing the number of API
requests, only the frequency with which WordPress checks on the
need to run a new request.