:sparkles: Auto-update data on activate, uninstall, and user register
Create methods to handle automatically fetching and updating the
user metadata when the plugin is first activated (to populate the
full usermeta data for the site) and when a user is created (to
fetch that user's data). Also create a method to remove the data
set by the plugin on plugin uninstall (delete).

Also cleans up some of the instantiation and loading methods.