Replace endpoint property setter with class set properties method
Switch from using `set_endpoint_props` method to assign the API
URL and the users to check to setting the URL once in a class
`set_properties` method and setting the users as-needed inside
the methods that need them. This also allows us to remove the
`users` class property.

Also adds a dedicated method to format email addresses into
usernames instead of repeating the logic multiple times.