:card_file_box: Sanitize API a11y data and clean up
Consolidate the accessibility data sanitizing and formatting
functions into the fetch method to, among other things,
guarantee the expiration date is a DateTime object and the
'isCertified' value is a booleen instead of a string.

Normalizing this data means we can also simplify later get
and print methods to only get and print values formatted for
display, instead of having to go through layers of formatting
first.

This also includes a new get method to fetch the WSU Accessibilty
Training course URL from the user metadata (instead of hard
coding it into the class method).