:bug: Fix #5 return expiration dates whether user certified or not
The notifications for expired certification did not display as
expected. The problem was not in the time difference function,
which correctly calculates the difference between two dates
irrespective of past or future. The problem was in the methods
to get the time, which returned false if the user was not
certified -- so naturally it returned false when a user's
certification expires instead of returning the time difference.