:bug: Fix #6 Correct date diff calculation for grace period
The method that returns the amount of time remaining in the
30-day grace period new users have to pass accessibility
training was counting up, not down. The problem was that the diff
function was counting days from registration to the current day,
not the current day to the expiration date. Fixed by using
DateInterval to set an end date for the grace period and then
testing between "today" and "end".