Replace email-to-nid conversion with usermeta and email fallback
Replace the email-to-username method of generating NIDs with the
more specific `get_user_wsu_nid` method, which checks the user's
usermeta for a saved WSU NID value and falls back to the previous
email-to-username conversion method. This solves the problem of
accessing data for users with a valid WSU NID who use a different
email address to log in to WordPress.