:bug: Fix #10 Direct refresh action to admin home to prevent errors
When a user viewing the posts or pages screen in the admin area
selected the "Refresh" button in the accessibility status notice
the site returned a "The link you followed has expired" error. To
prevent this -- since the action works as expected on most other
admin pages -- change the button path from the default of the
current admin page to specifically navigate to the admin home page
before triggering the action handler.