:bug: Fix #17 Don't overwrite `was_certified` value on expiration
The `update_a11y_user_meta` function replaced the entire existing
a11y user meta array with whatever the WSU API returned. This
meant that when a user's certification expired the "was certified"
value we use to compare "expired" to "never certified" would be
deleted.

This change adds logic to merge the existing user a11y status with
the updated values returned by the API. Since the API does not have
a "was certified" key, it will never be replaced.