Fix #18 Don't modify a11y expiration date when certification expires
Previously, when a user's certification expired the WSU API changed
their certification expiration date to a generic "12/25/2000." We
would rather store the date the user's certification expired so we
can quickly tell how long it has been since it expired.

This change will check if the user "was certified" and is now
"not certified." If so it will replace the new (incorrect) expiration
date from the API with the existing expiration date in the user meta.