:sparkles: Close #23 Create a settings page and set up plugin options
Register a settings page a nav menu item for the plugin (as a subpage
of the general options in the main Settings nav menu). This creates
a settings option group for the plugin in the *_options table of the
WP database. The option group is an array of plugin options, for now
only including the API URL. The plugin will populate the options with
default values on activation only if those values don't already exist.